requirements.txt
asgiref==3.4.1
crispy==0.6.3
cycler==0.11.0
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
fabio==0.12.0
fonttools==4.28.3
h5py==3.6.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.2
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
setuptools-scm==6.3.2
silx==1.0.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
